+++ ci/version-overrides/1.21.8/forge/src/main/java/app/jyu/forge/mixin/LevelRendererMixin.java
@@ -1,0 +1,24 @@
+package app.jyu.forge.mixin;
+
+import app.jyu.common.render.WorldRenderContext;
+import app.jyu.forge.event.WorldRenderCallback;
+import com.mojang.blaze3d.buffers.GpuBufferSlice;
+import com.mojang.blaze3d.resource.GraphicsResourceAllocator;
+import com.mojang.blaze3d.systems.RenderSystem;
+import net.minecraft.client.Camera;
+import net.minecraft.client.DeltaTracker;
+import net.minecraft.client.renderer.LevelRenderer;
+import org.joml.Matrix4f;
+import org.joml.Vector4f;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+
+@Mixin(LevelRenderer.class)
+public abstract class LevelRendererMixin {
+	@Inject(method = "renderLevel", at = @At(value = "TAIL"))
+	private void onStartRenderLevel(GraphicsResourceAllocator allocator, DeltaTracker deltaTracker, boolean renderBlockOutline, Camera camera, Matrix4f frustumMatrix, Matrix4f projectionMatrix, GpuBufferSlice fog, Vector4f fogColor, boolean renderSky, CallbackInfo ci) {
+		WorldRenderCallback.fire(WorldRenderContext.of(RenderSystem.getModelViewMatrix(), projectionMatrix, deltaTracker.getGameTimeDeltaTicks(), camera));
+	}
+}